+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-api/src/main/java/org/apache/shardingsphere/authority/auth/builtin/DefaultAuthentication.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.authority.auth.builtin;
 
 import lombok.Getter;
 import org.apache.shardingsphere.authority.auth.Authentication;
-import org.apache.shardingsphere.authority.auth.model.privilege.ShardingSpherePrivilege;
+import org.apache.shardingsphere.authority.model.ShardingSpherePrivilege;
 import org.apache.shardingsphere.infra.metadata.user.Grantee;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
