+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-api/src/main/java/org/apache/shardingsphere/authority/model/PrivilegeType.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.authority.auth.model.privilege;
+package org.apache.shardingsphere.authority.model;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
