+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/java/org/apache/shardingsphere/authority/loader/builder/loader/dialect/MySQLPrivilegeLoader.java
@@ -17,11 +17,11 @@
 
 package org.apache.shardingsphere.authority.loader.builder.loader.dialect;
 
-import org.apache.shardingsphere.authority.auth.model.privilege.database.SchemaPrivilege;
-import org.apache.shardingsphere.authority.auth.model.privilege.database.TablePrivilege;
+import org.apache.shardingsphere.authority.model.database.SchemaPrivilege;
+import org.apache.shardingsphere.authority.model.database.TablePrivilege;
 import org.apache.shardingsphere.authority.loader.builder.loader.PrivilegeLoader;
-import org.apache.shardingsphere.authority.auth.model.privilege.PrivilegeType;
-import org.apache.shardingsphere.authority.auth.model.privilege.ShardingSpherePrivilege;
+import org.apache.shardingsphere.authority.model.PrivilegeType;
+import org.apache.shardingsphere.authority.model.ShardingSpherePrivilege;
 import org.apache.shardingsphere.infra.metadata.user.Grantee;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
