+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/test/java/org/apache/shardingsphere/authority/loader/builder/loader/dialect/PostgreSQLPrivilegeLoaderTest.java
@@ -17,9 +17,9 @@
 
 package org.apache.shardingsphere.authority.loader.builder.loader.dialect;
 
-import org.apache.shardingsphere.authority.auth.model.privilege.PrivilegeType;
-import org.apache.shardingsphere.authority.auth.model.privilege.ShardingSpherePrivilege;
-import org.apache.shardingsphere.authority.auth.model.privilege.database.SchemaPrivilege;
+import org.apache.shardingsphere.authority.model.PrivilegeType;
+import org.apache.shardingsphere.authority.model.ShardingSpherePrivilege;
+import org.apache.shardingsphere.authority.model.database.SchemaPrivilege;
 import org.apache.shardingsphere.authority.loader.builder.loader.PrivilegeLoader;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
